HRCPP-204 Removed MurmurHash2 algorithm and ConsistentHashV1
