+++ src/hotrod/impl/consistenthash/ConsistentHashFactory.cpp
@@ -1,6 +1,5 @@
 #include "hotrod/impl/protocol/HotRodConstants.h"
 #include "hotrod/impl/consistenthash/ConsistentHashFactory.h"
-#include "hotrod/impl/consistenthash/ConsistentHashV1.h"
 #include "hotrod/impl/consistenthash/ConsistentHashV2.h"
 #include "infinispan/hotrod/Configuration.h"
 
@@ -14,9 +13,6 @@ using infinispan::hotrod::protocol::HotRodConstants;
 std::shared_ptr<ConsistentHash> ConsistentHashFactory::newConsistentHash(uint8_t version) {
     std::shared_ptr<ConsistentHash> result;
     switch (version) {
-        case HotRodConstants::CONSISTENT_HASH_V1:
-            result.reset(new ConsistentHashV1());
-            break;
         case HotRodConstants::CONSISTENT_HASH_V2:
             result.reset(new ConsistentHashV2());
             break;
