+++ src/hotrod/impl/protocol/HotRodConstants.h
@@ -20,7 +20,6 @@ class HotRodConstants
     static const uint8_t VERSION_12 = 12;
 
     //hotrod consistenthash version
-    static const uint8_t CONSISTENT_HASH_V1 = 1;
     static const uint8_t CONSISTENT_HASH_V2 = 2;
 
     // request operation code
